Cap stylistic sets complete I think
We should decide if we will have init and fina shapes and if we will: which ones. The presence of the serifed style mixed in will mean we have to sub from the init style back to the serifed style or that we keep the code for init and fina shapes limited in some way.